docker/baseline/layoutlm_parser.py
# Advanced Implementation: LayoutLM
# layoutlm_parser.py
from transformers import LayoutLMv3Processor, LayoutLMv3ForTokenClassification
from PIL import Image
import torch

class AdvancedDocumentParser:
    def __init__(self):
        self.processor = LayoutLMv3Processor.from_pretrained("microsoft/layoutlmv3-base")
        self.model = LayoutLMv3ForTokenClassification.from_pretrained("microsoft/layoutlmv3-base")

    def parse_with_layoutlm(self, image_path):
        """Parse document using LayoutLMv3"""
        image = Image.open(image_path).convert('RGB')

        # Process image
        encoding = self.processor(image, return_tensors="pt")

        # Get predictions
        with torch.no_grad():
            outputs = self.model(**encoding)
            predictions = outputs.logits.argmax(-1).squeeze().tolist()

        # Process predictions (simplified)
        tokens = self.processor.tokenizer.convert_ids_to_tokens(encoding.input_ids.squeeze().tolist())

        return {
            'tokens': tokens,
            'predictions': predictions,
            'processed': True
        }
